+++ mypyc/test/testutil.py
@@ -14,10 +14,10 @@ from mypy.test.data import DataSuite, DataDrivenTestCase
 from mypy.test.config import test_temp_dir
 from mypy.test.helpers import assert_string_arrays_equal
 
-from mypyc.genopsmain import build_ir
 from mypyc.options import CompilerOptions
 from mypyc.ops import FuncIR
 from mypyc.errors import Errors
+from mypyc.irbuild.main import build_ir
 from mypyc.irbuild.mapper import Mapper
 from mypyc.test.config import test_data_prefix
